date bindings. tests are WIP due to breaking symbol on mac
Comments:

1. Created date bindings to browser Date object
2. Some API changes:
    I) any javascript function that takes more than one variable (let's
call the function "variadic"), e.g. ```Date.utc()```, differs from the browser
API in that instead of taking several optional strings representing
various date constructs, we use an array of ```String```s here. This is
because Rust does not support optional arguments (as of 12/9/2017). One
way to have implemented this would be to use macros that could expand on
any given number of arguments, but this would cause the Date API to
become clumsy. A developer/consumer would have to know when a Date
method is a macro and when it is not (e.g. when to call ```Date::utc()``` vs
```date_utc!()```). Further, exporting macros are an
experimental feature in Rust at the moment and the current lib.rs
configuration is set up to block unstable features, for good reason. But
worst of all, making what should be an instance method a macro lead to
significant challenges. So, the solution here was to pass in an array representing the normal
arguments for the Date module as represented in the browser.
    II) types may be confusing at the moment. In an effort to keep
memory footprint low, I've been specific in laying out types for each
function (e.g. if we're returning a month number, f64 is huge for a
value between 1 and 12). That said, pretty much everything in javascript
is represented as a float, but with emscripten in the mix, it would be
great to be able to optimize some of these calls, before they make it to
the javascript world, as an extension of the C++ source of emscripten.
The whole point of being able to compile to (W)ASM is performance! I
hope the types are appopriate, both in the above reasoning and to
reflect current browser usage (the Date API)
    III) the string methods ```to_locale_date_string```,
```to_locale_string```, and ```to_locale_time_string``` all take two
(optional) arguments: locales and options. Instead of keeping with (I)
and making this an array API (e.g. ```to_locale_date_string([locales,
options])``` or ```to_locale_date_string()``` or
```to_locale_date_string([locales])```), I've made this require an API as ```to_locale_date_string(Option< &str >, Option< &HashMap < String, String > >)``` so a developer/consumer will need to provide "None" if options and/or locale is not used
    IV) In relation to (III), at the moment, sending locales as an array
is not supported (TODO)

3. tests are not working due to

Undefined symbols for architecture x86_64:
            "_emscripten_asm_const_int", referenced from:
                stdweb::webcore::value::Reference::from_raw_unchecked::hdbe203e0bb5dc861
in stdweb-d4848b03526ad141.stdweb1.rust-cgu.o
                _$LT$stdweb..webcore..value..Reference$u20$as$u20$core..ops..drop..Drop$GT$::drop::hd165fb3c2d7a11e6
in stdweb-d4848b03526ad141.stdweb1.rust-cgu.o
                stdweb::webcore::serialization::test_reserialization::i32::h18b05559d6914218
in stdweb-d4848b03526ad141.stdweb11.rust-cgu.o
                stdweb::webcore::serialization::test_reserialization::f64::hc629b32196a9e3cd
in stdweb-d4848b03526ad141.stdweb11.rust-cgu.o
                stdweb::webcore::serialization::test_reserialization::bool_true::h9c3b07d7362c2e53
in stdweb-d4848b03526ad141.stdweb11.rust-cgu.o
                stdweb::webcore::serialization::test_reserialization::bool_false::h677c538fa5c59948
in stdweb-d4848b03526ad141.stdweb11.rust-cgu.o
                stdweb::webcore::serialization::test_reserialization::undefined::hb9cfb96ef38391f7
in stdweb-d4848b03526ad141.stdweb11.rust-cgu.o
                ...
          ld: symbol(s) not found for architecture x86_64
          clang: error: linker command failed with exit code 1 (use -v
to see invocation)

4. For the functions taking optional arguments, they could definitely be
cleaned up to not need to expand arguments into a ```JsSerializable``` ```HashMap```.
However, there is an issue with the ```js!``` macro; any argument after the fourth throws an expansion error (I can provide more detail if need be, but it takes place around line 287 of macros.rs). Further, there is also an issue with the ```em_asm_double!``` (and presumably, any other ````em_asm_<data_type>!``` macro) where ```@{}``` syntax breaks on ```self```

5. The CSS I added could be refactored/changed as need be
6. I slightly changed the example TodoApp to use the new Date API. It
would be cool to add a "realtime" timer to the app (TODO).
7. This Date API could be used in the future by libraries built on top
of this, a la MomentJs, say