Clean up `abort`, `load` and `error` events
Some events (like these three) have multiple different events
assigned to the same name depending on what exactly they
are emitted by. For example, the `load` event is generated
by the DOM to signify that an element has finished loading,
and it's also sent by the `XMLHttpRequest` to tell us that
a given request has finished. In the first case the event
is derived from `UIEvent`, while in the second case the event
is based on `ProgressEvent`.

Two different events with the same name!

This is very confusing. Even people on MDN mix them up
and link to the wrong event on some pages!