Use Mut to let IEventTarget and MutationObserver accept FnMut again.
Without the changes to the runtime in the previous commit, letting these
functions accept FnMut can only be implemented with extra overhead. But
with these changes and the new Mut modifier, accepting FnMuts does not
incur extra overhead compared to Fn, because Fn also needs to check if
any calls are ongoing due to issue #278.