Fix compilation of one of the tests on older Rust
The test in question has a closure which has an `unreachable!` at the end,
which triggers the following error:

  "code relies on type inference rules which are likely to change"

This is most likely due to the fact that the result of `unreachable!` often
acts like a bottom type during the type-inference. To fix this we just
explicitly specify that the closure is supposed to return `()`.