Do not use custom sections after all
From what I can see the way Rust currently decides whenever to emit
a custom section is broken, which makes it impossible to force it
to do so. This in certain cases will result in JS shims which
generate an import but which are missing from the custom section.

So we'll temporarily go back to the way we previously used to emit
JS snippets, except this time we're using procedural macros.